requirements.txt
fastapi[all]
uvicorn
sqlalchemy
python-jose[cryptography]==3.5.0
pwdlib[argon2]